The MyAwesomePackage Module

MyAwesomePackage — Module

module MyAwesomePackage

API for doing awesome things.

More details about MyAwesomePackage

Module Index

MyAwesomePackage.plusOne

MyAwesomePackage.plusTwo

Detailed API

MyAwesomePackage.plusOne — Method

plusOne(x)

Sum the numeric "1" to whatever it receives as input.

A more detailed explaination can go here.

Arguments

x: The amount to which we want to add 1

Notes

Notes can go here

Examples

julia> five = plusOne(3)
4

MyAwesomePackage.plusTwo — Method

plusTwo(x)

Sum the numeric "2" to whatever it receives as input.

A more detailed explaination can go here.

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3)
5