The MyAwesomePackage Module

MyAwesomePackage — Module

MyAwesomePackage

Here you are, you found an awesome package :-)

Module Index

MyAwesomePackage.plusTwo

Detailed API

MyAwesomePackage.plusTwo — Method

plusTwo(x)

Sum the numeric "2" to whatever it receives as input

A more detailed explanation can goes here, altought I guess it is not needed in this case

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3)
5